Unify the host side code of new kds flow (#4711)
