+++ src/runtime_src/core/pcie/driver/linux/xocl/userpf/Makefile
@@ -88,7 +88,6 @@ xocl-y := \
 	../subdev/version_ctrl.o \
 	../subdev/msix_xdma.o \
 	../subdev/ert_user.o \
-	../subdev/ert_30.o \
 	../subdev/ert_versal.o \
 	$(xocl_lib-y)	\
 	$(drv_common-y) \
