+++ src/runtime_src/core/pcie/driver/linux/xocl/userpf/xocl_drv.c
@@ -1742,7 +1742,6 @@ static int (*xocl_drv_reg_funcs[])(void) __initdata = {
 	xocl_init_lapc,
 	xocl_init_msix_xdma,
 	xocl_init_ert_user,
-	xocl_init_ert_30,
 	xocl_init_ert_versal,
 	xocl_init_m2m,
 };
@@ -1780,7 +1779,6 @@ static void (*xocl_drv_unreg_funcs[])(void) = {
 	xocl_fini_lapc,
 	xocl_fini_msix_xdma,
 	xocl_fini_ert_user,
-	xocl_fini_ert_30,
 	xocl_fini_ert_versal,
 	xocl_fini_m2m,
 	/* Remove intc sub-device after CU/ERT sub-devices */
