+++ src/runtime_src/core/pcie/driver/linux/xocl/userpf/xocl_sysfs.c
@@ -290,11 +290,11 @@ kds_interrupt_store(struct device *dev, struct device_attribute *da,
 		goto done;
 
 	if (cu_intr) {
-		xocl_ert_30_mb_sleep(xdev);
-		xocl_ert_30_cu_intr_cfg(xdev);
+		xocl_ert_user_mb_sleep(xdev);
+		xocl_ert_user_cu_intr_cfg(xdev);
 	} else {
-		xocl_ert_30_mb_wakeup(xdev);
-		xocl_ert_30_ert_intr_cfg(xdev);
+		xocl_ert_user_mb_wakeup(xdev);
+		xocl_ert_user_ert_intr_cfg(xdev);
 	}
 
 	kds->cu_intr = cu_intr;
