+++ src/runtime_src/core/pcie/driver/linux/xocl/xocl_fdt.c
@@ -429,25 +429,11 @@ static struct xocl_subdev_map subdev_map[] = {
 		.dev_name = XOCL_ERT_USER,
 		.res_array = (struct xocl_subdev_res[]) {
 			{.res_name = NODE_ERT_CQ_USER, .regmap_name = PROP_ERT_CQ},
+			{.res_name = NODE_ERT_CFG_GPIO},
 			{.res_name = NODE_ERT_CQ_USER, .regmap_name = PROP_ERT_LEGACY},
 			{NULL},
 		},
 		.required_ip = 1,
-		.flags = XOCL_SUBDEV_MAP_USERPF_ONLY,
-		.build_priv_data = ert_build_priv,
-		.devinfo_cb = NULL,
-		.max_level = XOCL_SUBDEV_LEVEL_PRP,
- 	},
-	{
-		.id = XOCL_SUBDEV_ERT_30,
-		.dev_name = XOCL_ERT_30,
-		.res_array = (struct xocl_subdev_res[]) {
-			{.res_name = NODE_ERT_CFG_GPIO},
-			{.res_name = NODE_ERT_CQ_USER, .regmap_name = PROP_ERT_CQ},
-			{.res_name = NODE_ERT_CQ_USER, .regmap_name = PROP_ERT_LEGACY},
-			{NULL},
-		},
-		.required_ip = 2,
 		.flags = XOCL_SUBDEV_MAP_USERPF_ONLY,
 		.build_priv_data = ert_build_priv,
 		.devinfo_cb = NULL,
